requirements_rag_modern.txt
# Sistema RAG Moderno 2024 - Dependências
# Instalar com: pip install -r requirements_rag_modern.txt

# Core dependencies
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0

# LangChain ecosystem (recomendado)
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
langchain-huggingface>=0.0.1

# Vector databases
chromadb>=0.4.22
qdrant-client>=1.7.0

# Embeddings
sentence-transformers>=2.2.2
transformers>=4.36.0
torch>=2.1.0

# Document loaders
pypdf>=3.17.0
python-docx>=0.8.11
openpyxl>=3.1.2
beautifulsoup4>=4.12.0
lxml>=4.9.3
html2text>=2020.1.16
markdown>=3.5.0

# Text processing
tiktoken>=0.5.2
nltk>=3.8.1
spacy>=3.7.0

# API clients
openai>=1.3.0
anthropic>=0.8.0
google-generativeai>=0.3.0

# Web scraping (para integração futura)
selenium>=4.15.0
playwright>=1.40.0

# Utilities
tqdm>=4.66.0
colorama>=0.4.6
rich>=13.7.0
pydantic>=2.5.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0

# Optional: FastAPI for API server
fastapi>=0.104.0
uvicorn>=0.24.0

# Optional: Streamlit for web interface
streamlit>=1.28.0

# Optional: Jupyter for notebooks
jupyter>=1.0.0
ipykernel>=6.26.0

# Optional: Advanced embeddings
faiss-cpu>=1.7.4
# faiss-gpu>=1.7.4  # Uncomment for GPU support

# Optional: Database connectors
psycopg2-binary>=2.9.9  # PostgreSQL
pymongo>=4.6.0  # MongoDB
sqlalchemy>=2.0.23  # SQL databases

# Optional: Cloud storage
boto3>=1.34.0  # AWS S3
google-cloud-storage>=2.10.0  # Google Cloud Storage
azure-storage-blob>=12.19.0  # Azure Blob Storage

# Optional: Monitoring and logging
wandb>=0.16.0  # Weights & Biases
mlflow>=2.8.0  # MLflow
loguru>=0.7.2  # Advanced logging

# Optional: Performance
numba>=0.58.0  # JIT compilation
cython>=3.0.6  # C extensions